requirements.txt
absl-py==2.2.2
alembic==1.15.2
arrow==1.3.0
astunparse==1.6.3
attrs==24.2.0
bravado==11.0.3
bravado-core==6.1.1
catboost==1.2.8
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.55.1
fqdn==1.5.1
gapstatistics==0.1.5
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_resources==6.4.5
isoduration==20.11.0
joblib==1.4.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml_dtypes==0.5.1
monotonic==1.6
msgpack==1.1.0
namex==0.0.9
numba==0.60.0
numpy==2.0.2
nvidia-nccl-cu12==2.26.2.post1
opt_einsum==3.4.0
optree==0.15.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
protobuf==5.29.4
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shap==0.47.2
simplejson==3.19.3
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.40
swagger-spec-validator==3.0.4
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.5.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.3
webcolors==24.11.1
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
